+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.4.3
+- Added plugin includes caching
+- Added `cache clean` command
+
 # 1.4.2
 - Fixed project config generation